adk-agents/requirements.txt
google-cloud-aiplatform[adk,agent_engines]
requests
google-adk
pydantic
pydantic_settings
cloudpickle
